requirements.txt
streamlit
tensorflow
opencv-python
pillow
pandas
numpy
scikit-learn
matplotlib